pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alara"
version = "0.1.0"
description = "Ambient Language and Reasoning Assistant"
requires-python = ">=3.11"
dependencies = [
  "google-genai",
  "google-api-core",
  "prompt_toolkit>=3.0",
  "keyring",
  "cryptography",
  "anyio>=4.0",
  "mcp>=1.0",
  "httpx",
  "rich>=13.0",
  "tenacity>=8.0",
  "composio>=0.7",
]

[project.scripts]
alara = "alara.main:run"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
